Did streatch goal
I created the gradeRandom method which takes a student object as its parameter it then creates a points object which takes the floor of a random number from 0-100 because sometimes the number is like 88.5435435 and we dont need that. Then i make a bool based of a random number true return if its .5 or higher false if vice vera. I then created a graduate loop which checks if the grade is above 70 if so console logs a message and if not returns false i use this in two while true loops one for student sam and one for studen nathan and then randomly grade them till they return true and break out of the loop